Move drilldown component files to separate directory, add additional component
